requirements.txt
tensorflow~=2.1.0
tensorflow-federated==0.13.1
tf-encrypted-primitives==0.0.1